Move the client creating and validating functions to a more suitable package.
