+++ testsupport/space/spacerequest.go
@@ -1,7 +1,6 @@
 package space
 
 import (
-	"context"
 	"testing"
 
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/api/v1alpha1"
@@ -9,14 +8,9 @@ import (
 	"github.com/codeready-toolchain/toolchain-e2e/testsupport/util"
 	"github.com/codeready-toolchain/toolchain-e2e/testsupport/wait"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/client-go/tools/clientcmd/api"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -98,7 +92,7 @@ func NewSpaceRequest(t *testing.T, opts ...SpaceRequestOption) *toolchainv1alpha
 func VerifyNamespaceAccessForSpaceRequest(t *testing.T, cl client.Client, spaceRequest *toolchainv1alpha1.SpaceRequest) {
 	for _, nsAccess := range spaceRequest.Status.NamespaceAccess {
 		// create a kube client by ready the secret created in the spacerequest namespace
-		namespaceAccessClient, adminSecret := NewKubeClientFromSecret(t, cl, nsAccess.SecretRef, spaceRequest.Namespace)
+		namespaceAccessClient, adminSecret := util.NewKubeClientFromSecret(t, cl, nsAccess.SecretRef, spaceRequest.Namespace)
 
 		// check expected labels on the secret
 		require.NotEmpty(t, adminSecret.Labels[toolchainv1alpha1.SpaceRequestLabelKey])
@@ -107,44 +101,6 @@ func VerifyNamespaceAccessForSpaceRequest(t *testing.T, cl client.Client, spaceR
 		require.Equal(t, adminSecret.Labels[toolchainv1alpha1.SpaceRequestProvisionedNamespaceLabelKey], nsAccess.Name)
 
 		// validate the kube client has access to the namespace name that's in the spacerequest.Status.Namepsacess[n].Name field
-		ValidateKubeClient(t, namespaceAccessClient, nsAccess.Name, &corev1.SecretList{})
+		util.ValidateKubeClient(t, namespaceAccessClient, nsAccess.Name, &corev1.SecretList{})
 	}
-}
-
-// NewKubeClientFromSecret reads the kubeconfig from a given secret and create a kube rest client from it. You can supply functions to initialize
-// the scheme with which the client will be built.
-func NewKubeClientFromSecret(t *testing.T, cl client.Client, secretName, secretNamespace string, schemeAdders ...func(*runtime.Scheme) error) (client.Client, *corev1.Secret) {
-	t.Helper()
-	adminSecret := &corev1.Secret{}
-	// retrieve the secret containing the kubeconfig
-	require.NoError(t, cl.Get(context.TODO(), types.NamespacedName{
-		Namespace: secretNamespace,
-		Name:      secretName,
-	}, adminSecret))
-	assert.NotEmpty(t, adminSecret.Data["kubeconfig"])
-	apiConfig, err := clientcmd.Load(adminSecret.Data["kubeconfig"])
-	require.NoError(t, err)
-	require.False(t, api.IsConfigEmpty(apiConfig))
-
-	// create a new client with the given kubeconfig
-	kubeconfig, err := util.BuildKubernetesRESTConfig(*apiConfig)
-	require.NoError(t, err)
-
-	s := runtime.NewScheme()
-	builder := append(runtime.SchemeBuilder{}, corev1.AddToScheme)
-	builder = append(builder, schemeAdders...)
-	require.NoError(t, builder.AddToScheme(s))
-	namespaceAccessClient, err := client.New(kubeconfig, client.Options{
-		Scheme: s,
-	})
-	require.NoError(t, err)
-	return namespaceAccessClient, adminSecret
-}
-
-// ValidateKubeClient validates the the kube client can access the given namespace by listing objects using the provided list instance
-// The list is checked to not be empty.
-func ValidateKubeClient(t *testing.T, namespaceAccessClient client.Client, namespace string, list client.ObjectList) {
-	t.Helper()
-	require.NoError(t, namespaceAccessClient.List(context.TODO(), list, client.InNamespace(namespace)))
-	require.NotEmpty(t, list)
 }
